Feature/sg 344 refactor modules (#438)
* Added CSPResNet backbone

* Rename variables to make linter happy

* Rename typo in filename

* Rename arguments to be consistent with argument names in nn.Conv2d

* Introduce a general ConvBNAct to module_utils.py

* Fixed bug in not using activation_type

* Move ConvBNAct and RepVGG blocks to sg.modules

* Modules renaming for consistent naming
EffectiveSELayer -> EffectiveSEBlock
BasicBlock -> CSPResNetBasicBlock

* Added documentation

* Added documentation

* Move ConvBNReLU to sg.modules.__init__ temporary.
