+++ src/super_gradients/training/utils/module_utils.py
@@ -1,6 +1,6 @@
 from collections import OrderedDict
 import copy
-from typing import List, Union, Tuple, Optional
+from typing import List, Union, Optional
 import torch
 from torch import nn
 
@@ -45,7 +45,7 @@ class MultiOutputModule(nn.Module):
         """
         super().__init__()
         self.output_paths = output_paths
-        self._modules['0'] = module
+        self._modules["0"] = module
         self._outputs_lists = {}
 
         for path in output_paths:
@@ -61,7 +61,7 @@ class MultiOutputModule(nn.Module):
 
     def forward(self, x) -> list:
         self._outputs_lists[x.device] = []
-        self._modules['0'](x)
+        self._modules["0"](x)
         return self._outputs_lists[x.device]
 
     def _get_recursive(self, module: nn.Module, path) -> nn.Module:
@@ -100,10 +100,7 @@ class MultiOutputModule(nn.Module):
 
     def _slice_odict(self, odict: OrderedDict, start: int, end: int):
         """Slice an OrderedDict in the same logic list,tuple... are sliced"""
-        return OrderedDict([
-            (k, v) for (k, v) in odict.items()
-            if k in list(odict.keys())[start:end]
-        ])
+        return OrderedDict([(k, v) for (k, v) in odict.items() if k in list(odict.keys())[start:end]])
 
 
 def _replace_activations_recursive(module: nn.Module, new_activation: nn.Module, activations_to_replace: List[type]):
@@ -125,77 +122,26 @@ def replace_activations(module: nn.Module, new_activation: nn.Module, activation
     :param activations_to_replace:  types of activations to replace, each must be a subclass of nn.Module
     """
     # check arguments once before the recursion
-    assert isinstance(new_activation, nn.Module), 'new_activation should be nn.Module'
-    assert all([isinstance(t, type) and issubclass(t, nn.Module) for t in activations_to_replace]), \
-        'activations_to_replace should be types that are subclasses of nn.Module'
+    assert isinstance(new_activation, nn.Module), "new_activation should be nn.Module"
+    assert all(
+        [isinstance(t, type) and issubclass(t, nn.Module) for t in activations_to_replace]
+    ), "activations_to_replace should be types that are subclasses of nn.Module"
 
     # do the replacement
     _replace_activations_recursive(module, new_activation, activations_to_replace)
 
 
 def fuse_repvgg_blocks_residual_branches(model: nn.Module):
-    '''
+    """
     Call fuse_block_residual_branches for all repvgg blocks in the model
     :param model: torch.nn.Module with repvgg blocks. Doesn't have to be entirely consists of repvgg.
     :type model: torch.nn.Module
-    '''
+    """
     assert not model.training, "To fuse RepVGG block residual branches, model must be on eval mode"
     for module in model.modules():
-        if hasattr(module, 'fuse_block_residual_branches'):
+        if hasattr(module, "fuse_block_residual_branches"):
             module.fuse_block_residual_branches()
     model.build_residual_branches = False
-
-
-class ConvBNReLU(nn.Module):
-    """
-    Class for Convolution2d-Batchnorm2d-Relu layer. Default behaviour is Conv-BN-Relu. To exclude Batchnorm module use
-        `use_normalization=False`, to exclude Relu activation use `use_activation=False`.
-    For convolution arguments documentation see `nn.Conv2d`.
-    For batchnorm arguments documentation see `nn.BatchNorm2d`.
-    For relu arguments documentation see `nn.Relu`.
-    """
-
-    def __init__(self,
-                 in_channels: int,
-                 out_channels: int,
-                 kernel_size: Union[int, Tuple[int, int]],
-                 stride: Union[int, Tuple[int, int]] = 1,
-                 padding: Union[int, Tuple[int, int]] = 0,
-                 dilation: Union[int, Tuple[int, int]] = 1,
-                 groups: int = 1,
-                 bias: bool = True,
-                 padding_mode: str = 'zeros',
-                 use_normalization: bool = True,
-                 eps: float = 1e-5,
-                 momentum: float = 0.1,
-                 affine: bool = True,
-                 track_running_stats: bool = True,
-                 device=None,
-                 dtype=None,
-                 use_activation: bool = True,
-                 inplace: bool = False):
-
-        super(ConvBNReLU, self).__init__()
-        self.seq = nn.Sequential()
-        self.seq.add_module("conv", nn.Conv2d(in_channels,
-                                              out_channels,
-                                              kernel_size=kernel_size,
-                                              stride=stride,
-                                              padding=padding,
-                                              dilation=dilation,
-                                              groups=groups,
-                                              bias=bias,
-                                              padding_mode=padding_mode))
-
-        if use_normalization:
-            self.seq.add_module("bn", nn.BatchNorm2d(out_channels, eps=eps, momentum=momentum, affine=affine,
-                                                     track_running_stats=track_running_stats, device=device,
-                                                     dtype=dtype))
-        if use_activation:
-            self.seq.add_module("relu", nn.ReLU(inplace=inplace))
-
-    def forward(self, x):
-        return self.seq(x)
 
 
 class NormalizationAdapter(torch.nn.Module):
@@ -208,6 +154,7 @@ class NormalizationAdapter(torch.nn.Module):
      number of input channels.
 
     """
+
     def __init__(self, mean_original, std_original, mean_required, std_required):
         super(NormalizationAdapter, self).__init__()
         mean_original = torch.tensor(mean_original).unsqueeze(-1).unsqueeze(-1)
@@ -223,9 +170,7 @@ class NormalizationAdapter(torch.nn.Module):
         return x
 
 
-def make_upsample_module(scale_factor: int,
-                         upsample_mode: Union[str, UpsampleMode],
-                         align_corners: Optional[bool] = None):
+def make_upsample_module(scale_factor: int, upsample_mode: Union[str, UpsampleMode], align_corners: Optional[bool] = None):
     """
     Factory method for creating upsampling modules.
     :param scale_factor: upsample scale factor
